pyproject.toml
[project]
name = "slide-renderer"
version = "0.1.0"
description = "Render JSON slides to Marp markdown presentations with Jinja2 templates"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "B-Lab Theme Contributors", email = "blab@example.com" }
]
keywords = ["marp", "slides", "json", "markdown", "jinja2", "presentation", "llm", "renderer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup",
]
license = {text = "MIT"}

dependencies = [
    "jinja2>=3.0.0",
    "openai>=2.7.2",
    "pydantic>=2.0.0,<3.0.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-username/slide-renderer"
Repository = "https://github.com/your-username/slide-renderer"
Issues = "https://github.com/your-username/slide-renderer/issues"
Documentation = "https://github.com/your-username/slide-renderer#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/slide_renderer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=slide_renderer --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
